+++ lib/src/picker/model.dart
@@ -1,4 +1,5 @@
 import 'package:event_db/event_db.dart';
+import 'package:tuple/tuple.dart';
 
 class SelectedTheme extends GenericModel {
   late String name;